+++ .env_example
@@ -1,3 +1,3 @@
 DB_HOST=mongodb+srv://<DB_USER>:<DB_USER_PASS>@cluster0.ldebl.mongodb.net/<DB_NAME>?retryWrites=true&w=majority
 SECRET_KEY=
-SENDGRID_API_KEY
+SENDGRID_API_KEY=
